+++ src/chains/stellar/index.ts
@@ -1,4 +1,6 @@
 export { deriveStealthKeys, deriveStealthKeysFromSigner } from './keys';
+export type { KeyDerivationOptions } from './keys';
+export type { Tracer, Span } from '../../telemetry';
 export { FreighterStealthSigner, WebAuthnPasskeyStealthSigner } from './signer';
 export type {
   StellarStealthSigner,
@@ -22,7 +24,7 @@ export { generateStealthAddress } from './stealth';
  * @internal
  */
 export { computeSharedSecret, computeAnnouncementViewTag, computeViewTag } from './stealth';
-export { checkStealthAddress, scanAnnouncements, adapter as stellarAdapter, adapter } from './scan';
+export { checkStealthAddress, scanAnnouncements, scanAnnouncementsStream } from './scan';
 /**
  * @internal
  */
@@ -136,4 +138,4 @@ export { createHorizonClient } from './horizon';
 export type { RetryPolicy, HorizonClient, HorizonClientConfig } from './horizon';
 
 export { createRpcClient } from './rpc';
-export type { RpcClient, RpcClientConfig, RpcEndpoint } from './rpc';
+export type { RpcClient, RpcClientConfig, RpcEndpoint, RpcRequestOptions } from './rpc';